requirements.txt
python-dotenv
requests
pyyaml
SQLAlchemy==1.4.47
pandas
psycopg2